requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
pdfminer.six==20240706
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0
langdetect==1.0.9
gTTS==2.5.4
# Lightweight translation (no Rust required)
deep-translator==1.11.4
# Optional alternatives:
# pyttsx3==2.90
# coqpit==0.0.17
# TTS==0.22.0  # Coqui TTS (heavy)
# easyocr==1.7.1  # Alternative OCR (heavy)
